+++ sdk/keyvault/Azure.Security.KeyVault.Certificates/api/Azure.Security.KeyVault.Certificates.netstandard2.0.cs
@@ -295,6 +295,7 @@ namespace Azure.Security.KeyVault.Certificates
         public CertificateProperties(System.Uri id) { }
         public System.DateTimeOffset? CreatedOn { get { throw null; } }
         public bool? Enabled { get { throw null; } set { } }
+        public string EntityId { get { throw null; } }
         public System.DateTimeOffset? ExpiresOn { get { throw null; } }
         public System.Uri Id { get { throw null; } }
         public string Name { get { throw null; } }